using buffers to transfer AssetGraph to nodejs
