+++ crates/node-bindings/src/atlaspack/mod.rs
@@ -8,4 +8,5 @@ pub mod file_system_napi;
 pub mod monitoring;
 pub mod napi_result;
 pub mod package_manager_napi;
+pub mod serialize_asset_graph;
 pub mod worker;
